Add check if Nextjs build is already built

+++ frontend/entrypoint.sh
@@ -6,20 +6,22 @@
 # that we can build the frontend after an other service has started in docker compose
 # (That's why the entrypoint.sh is here)
 
-npm run build
+if [ -e "./standalone/server.js" ]; then
+    echo "Server.js already exists, skipping build process."
+else
+    cd ./app
+    npm run build
 
-rm -rf node_modules
+    # https://nextjs.org/docs/app/api-reference/next-config-js/output#automatically-copying-traced-files
+    rm -rf node_modules
 
-# https://nextjs.org/docs/app/api-reference/next-config-js/output#automatically-copying-traced-files
+    cp -r ./public ./.next/standalone/public
+    cp -r ./.next/static ./.next/standalone/.next/static
 
+    cp -r ./.next/standalone ../
 
-cp -r ./public ./.next/standalone/public
-cp -r ./.next/static ./.next/standalone/.next/static
-
-cp -r ./.next/standalone ../
-
-# clean up
-cd .. 
-rm -r ./app
+    cd ..
+    rm -r ./app
+fi
 
 exec node ./standalone/server.js
